backend/asr_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
grpcio==1.62.1
grpcio-tools==1.62.1